Remove TODO
The datasheet isn't clear about this, but since the min value is inclusive, I think the max should be, too. And either way, it's highly unlikely that a correct controller will get anywhere near violating this while properly implementing refreshes, so being one cycle off shouldn't be a big deal.