Add tREF tester and basic refresh logic
The way this is currently formulated makes checking for violations REALLY slow, so tREF is currently incorrectly specified as 1ms instead of 64ms. This is fine for testing but won't fly for a real impl (it's conservatively within spec, but will be extremely inefficient). I'd like to rethink the checking approach and see if it's possible to improve this, even for cases where all rows have been refreshed and need to be checked.